+++ desktop/src/main/kotlin/com/music/vivi/desktop/Navigation.kt
@@ -39,6 +39,7 @@ sealed interface Screen {
     data object SettingsAbout : Screen
     data object SettingsDeveloper : Screen
     data object SettingsBackup : Screen
+    data object SettingsNotifications : Screen
 }
 
 /**